feat: add functionality to copy text when clicking on a 'p' element
Implements a functionality that allows the user to copy the text inside a 'p' element when clicking on it. The functionality uses the Clipboard API of JavaScript to copy the text to the clipboard. When clicking on a 'p' element, the text is copied and a success message is displayed to the user.