style: change the icon's source
Previously, i tried to use the Bootstrap icon set but ran into issues with the icons not displaying correctly. After some investigation, i discovered that an event was being set to all elements with the class 'bi', which was causing the problem because all of bootstrap's icon, has the class 'bi'. To solve this, i added a class called 'navBtn' to each element that i wanted to have the event and updated the CSS accordingly. Now the icons are displaying correctly.